Update README.md
e are navigating co-ordinates of all the directions using gyroscope and code respective to this directions.  we will move our vehicle by moving gyroscope in specific co-ordinates. so basically the car will follow our hand gestures and move accordingly. also we have added a ultra sonic sensor for the object detection so that it can sense the objects ahead and stop without any pre-signal so that we can avoid the accidents. we can further add features like pot-hole detections and automatic voice control AI based vehicle in future.